cache fix,1 flip,wait_end; global quit; mock sound
Cache fixation TextStim so it doesn't have to be created within inner loop.
Though that is likely uncessary with addition of `wait_end`:
  compute ideal wait time before any trigger or screen flip.
  Then core.wait can always wait the correct amount of time -- no slips
  b/c some process takes longer (esp screen flip)

Reorganized sound! Should play same sounds on open or close.
One sound (633.wav) is to indicate start of block.
The other (Correct.wav) to indicate end (and okay to open eyes).

global quit key (escape) needed to be configured AFTER `win` is created.
Hitting escape now ends the task early.

Needed to mock sound on local laptop to continue.
Currently getting a local segfault that's not worth debugging.

timemark now caches it's last call to give differences.
This let me clean up a lot of the t=timemark instances.

Like the sound mock, it's a kludge that the linter doesn't like